fix(ios,android): polish wavy/dotted/dashed text decoration rendering
Iterations on the textDecorationStyle implementation that landed in PR
#56748, based on visual comparison with Chrome / Safari and side-by-side
testing of the two platforms.

iOS:
- Wavy: thickness divisor relaxed from `fontSize / 8` to `fontSize / 12`
  and control-point distance multiplier halved (`1.5 * thickness + 0.5`
  vs Blink's literal `3 * thickness + 0.5`). At iOS point sizes the
  literal Blink amplitude renders as a very pronounced wave; the dialed-
  back values read as a clear-but-subtle browser-style wave.
- Dotted: switched from UIKit's `NSUnderlineStylePatternDot` (which
  doesn't match browser geometry) to a custom CG path with a zero-length
  dash + round line caps, producing actual circular dots at `2 *
  thickness` spacing.
- Dashed: switched from UIKit's `NSUnderlineStylePatternDash` to
  custom CG path with `[2 * thickness, thickness]` intervals — short
  rectangular dashes with a tight gap, closer to Safari's geometry than
  UIKit's default.
- The custom decoration attribute (formerly `RCTWavyDecorationAttributeName`)
  is now `RCTCustomDecorationAttributeName` and carries a `style` key so
  the same drawing pipeline handles wavy + dotted + dashed.

Cross-platform:
- Wavy drawing loop now iterates `while x < x2` instead of
  `while x + wavelength <= x2`, so the final cycle continues through
  the last character (including trailing punctuation). Previously a
  trailing period could be visually uncovered when the run width was
  not an integer multiple of the wavelength.

## Changelog:

[IOS] [CHANGED] - Wavy, dotted, and dashed text decorations render with custom CoreGraphics paths instead of UIKit pattern bits, matching browser geometry more closely
[GENERAL] [FIXED] - Wavy underline / strikethrough now extends through the final character of the run, including trailing punctuation

## Test Plan:

Side-by-side comparison on Android API 36 emulator and iPhone 17 sim
(iOS 26.4) of a `<Text>` with `textDecorationLine="underline"` and
`textDecorationStyle` cycling through `wavy` / `dotted` / `dashed`,
verified against Chrome (Android view) and Safari (iOS view) rendering
of the same CSS. Trailing periods now fall under the wavy stroke on
both platforms.
