+++ packages/react-native/ReactAndroid/src/main/java/com/facebook/react/views/text/internal/span/ReactStrikethroughSpan.kt
@@ -9,25 +9,19 @@ package com.facebook.react.views.text.internal.span
 
 import android.graphics.Canvas
 import android.graphics.Color
-import android.os.Build
 import android.text.Layout
 import com.facebook.react.views.text.TextDecorationStyle
-import com.facebook.react.views.text.drawDecorationLine
-import kotlin.math.max
+import com.facebook.react.views.text.drawSpannedDecoration
 
 /**
- * Draws a strikethrough whose color may differ from the text color and
- * whose stroke style may be `solid`, `double`, `dotted`, or `dashed`.
- * Subclasses [DrawCommandSpan] so [PreparedLayoutTextView] and
- * [ReactTextView] invoke [onDraw] after the layout renders its text. We
+ * Draws a strikethrough whose color and style may differ from the text.
+ * The line is painted in `onDraw` after the layout renders its text. We
  * do NOT extend [android.text.style.StrikethroughSpan] here: the
  * framework's `Layout.draw` paints the strikethrough using `paint.color`
- * with no field to override, so the only way to get a distinct color (or
- * style) is to draw it ourselves.
+ * with no field to override, so painting it ourselves is the only way to
+ * get a distinct color or non-solid style.
  *
- * When [color] is [Color.TRANSPARENT] (the default when no
- * `textDecorationColor` prop was passed), the strikethrough is drawn in
- * the text's foreground color, matching the platform's prior behavior.
+ * `color == Color.TRANSPARENT` falls back to the text foreground color.
  */
 internal class ReactStrikethroughSpan(
     private val color: Int = Color.TRANSPARENT,
@@ -35,52 +29,13 @@ internal class ReactStrikethroughSpan(
 ) : DrawCommandSpan() {
 
   override fun onDraw(start: Int, end: Int, canvas: Canvas, layout: Layout) {
-    val paint = layout.paint
-    val savedColor = paint.color
-    val savedStrokeWidth = paint.strokeWidth
-    val savedStyle = paint.style
-    val savedAntiAlias = paint.isAntiAlias
-    val effectiveColor = if (color != Color.TRANSPARENT) color else savedColor
-    // Density-aware minimum so the strikethrough reads consistently
-    // across display densities. `paint.density` is the px-per-dp ratio
-    // at the current paint setup, so `1.5f * paint.density` gives ~1.5 dp.
-    val minThickness = 1.5f * paint.density
-    val thickness =
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-          max(paint.underlineThickness, minThickness)
-        } else {
-          max(paint.fontMetrics.descent * 0.1f, minThickness)
-        }
-
-    paint.color = effectiveColor
-    paint.strokeWidth = thickness
-    paint.style = android.graphics.Paint.Style.STROKE
-    paint.isAntiAlias = true
-
-    // Position the strikethrough slightly below the midpoint between
-    // the line's top and baseline so it sits near the x-height midline
-    // like the platform default. `fontMetrics.ascent` is negative and
-    // `descent` is positive, so the sum / 2 gives a small negative
-    // offset from the baseline; the trailing `+ 1f` nudges it down to
-    // match the visual position users expect.
-    val fm = paint.fontMetrics
-    val offset = (fm.ascent + fm.descent) / 2f + 1f
-
-    val startLine = layout.getLineForOffset(start)
-    val endLine = layout.getLineForOffset(end)
-    for (line in startLine..endLine) {
-      val baseline = layout.getLineBaseline(line).toFloat()
-      val x1 =
-          if (line == startLine) layout.getPrimaryHorizontal(start) else layout.getLineLeft(line)
-      val x2 =
-          if (line == endLine) layout.getPrimaryHorizontal(end) else layout.getLineRight(line)
-      val y = baseline + offset
-      drawDecorationLine(canvas, paint, x1, x2, y, thickness, style)
+    drawSpannedDecoration(start, end, canvas, layout, color, style) { paint, baseline, _ ->
+      // Strikethrough sits near the x-height midline. `fontMetrics.ascent`
+      // is negative and `descent` is positive, so the sum / 2 gives a
+      // small negative offset from the baseline; the trailing `+ 1f`
+      // nudges it down to match the visual position users expect.
+      val fm = paint.fontMetrics
+      baseline + (fm.ascent + fm.descent) / 2f + 1f
     }
-
-    paint.color = savedColor
-    paint.strokeWidth = savedStrokeWidth
-    paint.style = savedStyle
-    paint.isAntiAlias = savedAntiAlias
   }
 }
